fix: hide unsupported loop commands
